Random Direction model
In the Random Direction model, in contrast, each node chooses a direction uniformly at
random (from the set of directions that would keep it within the field) and follows it with a
speed  chosen  uniformly  at  random  from  the  interval  [Vmin;  Vmax]  until  it  reaches  a 
boundary.  Then, it pauses for T time steps. It repeats the same process over and 
over.